[CHANGE] Move `staffview` class to `_staffview.py`
